vtk: reference-nsih: modify the pll-ddr0 frequency
This patch changes the frequency of pll-ddr0 to half for boards
that could not operate reliably at high speeds. The memory timing
is 800Mhz, but since the actual clock is 400Mhz, the memory speed
can be slower. This patch will be applied temporarily until a stable
memory calibration code is patched.

Change-Id: Ic1b6db96782edb31c49e4a4aa9f1120723e4cc60
Signed-off-by: DeokJin Lee <truevirtue@nexell.co.kr>